refactor: 컴포넌트 위치 수정 및 절대경로 적용
